// apps/admin/src/shared/api/generated/site/schemas/footerMenuResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — Site subset
 * Filtered for Homepage Admin site Orval codegen.
 * OpenAPI spec version: v1
 */
import type { FooterMenuResponseLinkType } from './footerMenuResponseLinkType';

export interface FooterMenuResponse {
  menuCode?: string;
  label?: string;
  linkType?: FooterMenuResponseLinkType;
  internalRouteCode?: string;
  externalUrl?: string;
  enabled?: boolean;
  displayOrder?: number;
  version?: number;
}